(Initial) Updating original plugin by @aviskase
This is the initial commit updating the code by @aviskase for the "Link Indexer" plugin for Obsidian. I've added the ability to alphabetize the list of links, placed the list into a table for better viewing and better export into an HTML page, and also added how other terms that connect to the main term.