Fixing the Graph Links
The Graph Links are still not working. I did more digging and research with the help of a research AI and found out that it appears that direct Graph View linking using the URI scheme isn't officially supported in Obsidian. This explains why these approaches aren't working despite seeming logical.

Since I'm building a plugin (not just creating Markdown), I do have more options available. So, I am going to implement a solution that actually works this time. Though, still is buggy when outputted to HTML using the Quartz system.

So, the redesign in the approach is to use Obsidian's API directly.

The key thing to understand is that we can't create direct links to the graph view in the Markdown content -- yet. Instead, we need to create clickable elements that trigger JavaScript code to perform the desired navigation.

The result, if there is a note that has been created for the term, it opens along side a Graph View of that term showing how it is connected throughout the vault!